openstack: use existing floating IP for LB
The OpenStack provider now does not create any floating IP addresses.
Instead, the user is expected to create one themselves ahead of time and
pass it in.

This lets us support non-FIP deployments and it is the final step
towards a completely unattended installation.

Since the external API DNS records have to be created out-of-band and
since the floating IP address could not be determined ahead of time,
this was something the deployers had to do as the installation was
happening.

Now they can create the FIP, create the DNS records, add the IP to the
install config (under `platform.openstack.lbFloatingIP`) and have the
start the installation and let it complete on its own.

If the `lbFloatingIP` value is not specified, no floating IP will be
created. This will likely make the installer fail (unless the deployer
configured their networking and DNS such that the API is externally
accessible), but this was true even before this change.

This should still result in a successful deployment, the installer just
won't be able to tell.