Display CVO progress during installs
As a command line application which is the primary interface
for developers, I think the installer should have more information
than it does today by default.  The CVO status is *extremely*
useful and informative - let's start showing it.

Digging through the log files in debug mode afterwards is too
annoying.

This doesn't add a significant amount of new logging, and
there are higher level wrappers for the installer including
a new web site that should probably be the first point of contact
for anyone who wants their hand held more.