types: allow compute pools with names other than 'worker'
Relax the restriction that there is exactly one compute machine pool and that
it is named "worker". Also, relax the restriction that a compute machine pool
must have a positive number of replicas.

Warn the user if the install-config.yaml does not specify any compute
nodes. The user will not be blocked from using no compute nodes.

Changes for https://jira.coreos.com/browse/CORS-988