manifest: set the public and private zones for AWS
The `DNS` object [2] was expanded to include the pointers to the public and private hosted zones to allow operators to create DNS records inthe correct zone in PR #202 [4].

The public zone is fetched from the `basedomain` while since installer creates the private zone at later stage, tags are used to specify the private dns zone.

The ID returned by `get-hosted-zone` [1] is of the form `/hostedzone/<id>`. For example,

```console
$ AWS_PROFILE=openshift-dev aws route53 get-hosted-zone --id XXX
{
    "HostedZone": {
        "Id": "/hostedzone/XXX",
...
}
```

This creates the `DNS` [2] object like:
```yaml
apiVersion: config.openshift.io/v1
kind: DNS
metadata:
  name: cluster
spec:
  baseDomain: yyy.openshift.com
  privateZone:
    tags:
      Name: adahiya-1_int
      kubernetes.io/cluster/adahiya-1: owned
      openshiftClusterID: 25375c13-3c0a-4102-a8f9-7ecb60757c62
  publicZone:
    id: "/hostedzone/XXX"
status: {}
```

But you can actually use both `/hostedzone/XXX` or `XXX` to get a zone.

```console
$ AWS_PROFILE=openshift-dev aws route53 get-hosted-zone --id "/hostedzone/XXX"
{
    "HostedZone": {
        "Id": "/hostedzone/XXX",
...
}
```

The change trims the `zoneID` to not include the prefix when creating the `DNS` object as just the ID has cleaner semantics when stored in the object. The terraform-provider-aws also cleans the zone-id for better UX [3] by trimming the prefix `/hostedzone/`.

[1]: https://docs.aws.amazon.com/cli/latest/reference/route53/get-hosted-zone.html#synopsis
[2]: https://github.com/openshift/api/blob/8c839bc7ff62e38ad7656bf920e11e2664a44f6b/config/v1/types_dns.go#L11
[3]: https://github.com/terraform-providers/terraform-provider-aws/blob/75a9ebb7bfc7fd61b4454589155cc6b958ebebe4/aws/resource_aws_route53_zone.go#L461-L464