types: validate cluster name in InstallConfig
Validate that the cluster name in an InstallConfig is a valid DNS subdomain.

Validate that <cluster-name>.<base-domain> is a valid DNS subdomain.

The DomainName function in pkg/validate now takes a parameter specifying
whether the domain name being validated is permitted to have a trailing dot.
The base domain is allowed to have a trailing dot, but the cluster name is
not allowed to have a trailing dot.

Fix for https://bugzilla.redhat.com/show_bug.cgi?id=1663447