openshift-install: add version string to logs
Including the version in the logs should make debugging a little easier
since we'll definitively know which version of the installer produced
them. This log entry was added under `setupFileHook()` since it felt
natural to include it as part of the logfile setup: ensure the logging
directory exists, ensure the logfile exists, append the current version.