image: Take explicit dependency on util-linux for uuidgen and gzip
uuidgen is moderately useful to a consumer to generate a unique
cluster id, but we are in the process of removing util-linux which
contains uuidgen from the base image because it brings in about
30M of compressed dependencies above what the new minimal base
includes (timezone data, systemd libs, a few others). For now,
update the image to make the dependency explicit, although we
may want to remove uuidgen in the future or make it a subcommand
of the install.

Gzip is used to compress artifacts in CI and is also moderately
useful for an admin who wants to compress logs.