pkg/destroy: data/aws: delete undiscoverable AWS objects
Today we find instance profiles that we need to delete through either the
associated role or the associated instance. We find those objects via aws
tags. If those objects have been deleted we are unable to find the instance
profiles because those are not tagged.

Since we embed both the cluster id inside the name we can find the instance
profiles we created by name and destroy them that way.