types/validation: test validation of non-empty, invalid CIDRs
The tests that validate CIDRs in install-config.yaml were only testing
empty CIDRs. These changes test non-empty but still invalid CIDRs instead.

The ipnet.ParseCIDR function was changed to use the IP as read from the
CIDR string instead of the IP from the parsed CIDR. This makes the
function consistent with the UnmarshalJSON function.