*: increase resources for libvirt machines
OpenShift has picked up a little weight over the past few months. Two
and four gibibytes of RAM for the control plane and compute machines,
respectively, is no longer enough. This bumps the machines to four and
six gibibytes. It also bumps the vCPUs on the masters from two to four.