uninstaller: fix endless loop when no public route53 zone found
When the uninstaller no longer finds an associated public zone
(because the records associating the public zone to the private zone no longer exist),
it currently loops endlessly because the query to retrieve records of the public zone
always fails. It never reaches the part that deletes the private zone.

This fix checks that the public zone id is not empty in order to query
its records.