data/aws: split API dns creation from etcd dns creation
We need the API DNS records in order for ignition to run on the masters. But
today we need the master's IP addresses to create the etcd DNS records. This PR
splits the creation of the API DNS record (which doesn't need to know the master
IPs) from the etcd records. In this way the api dns records are created much
sooner.