data/aws: place bootstrap node in same az as first master
The bootstrap node was being placed in the first availability zone in the region.
Now, place the bootstrap node in the same availability zone as the first master.

Remove the local az_to_private_subnet_id variable from the vpc module
as it is only used as an output from the module. The output value is now calculated
at the place where the output value is defined.

Remove the cluster_id output value from the vpc module as it is unused.