docs: add details on install time customization for machine configuration
The docs include an example for removing `NoSchedule` taint on masters to allow for single master cluster. Although the changes described in the
example are not completely sufficient to achieve the single master cluster, as few operators specifcally set workloads to worker nodes.

I chose this example because the Code Ready Containers team will be using somehting similar to achieve single node libvirt cluster.