destroy: delete snapshots used in ami copies
When the encrypted AMI used by the cluster is created, a snapshot
is created as well. This snapshot needs to be explicitly deleted
when the cluster is destroyed to avoid leaking resources. Unfortunately,
the snapshot does not have any tags when it is created. The snapshot
can be found from the AMI's block device mappings, though. When the
destroyer finds an AMI to delete, the destroyer will tag the snapshots
used by the AMI with tags that match the filter used to find the AMI.
This will allow the destroyer to find the snapshots on subsequent queries
using that filter.

It would be preferably if the snapshot were tagged during the cluster
creation. There is no way that I see to tag the snapshot when the AMI is
created, using the terraform aws provider. There may be a way to
have terraform assume ownership of the snapshot after the AMI is created
and modify the tags of the snapshot after that. Or there may be a way to
run some shell code from within terraform to modify the tags (although
that would necessitate having the AWS CLI installed, which I think is a
non-starter). In any event, those are things that could be looked into
as a follow-on. But for now, the destroyer needs to stop leaking the
snapshots, and these changes accomplish that.

Fixes https://github.com/openshift/installer/issues/1342