*: Wrapf(err, "...") -> Wrap(err, "...")
No need for formatting in these cases.  I've also adjusted the "Could
not {attempted thing}" message to be more consistent with our other
"failed to {attempted thing}" messages.

Personally I prefer "{attempted thing}", because the fact that we're
returning an error sufficiently expresses our failure.  But that's
been occasionally contentious before, so I'm punting on it while
touching these lines ;).