data/aws: create bootstrap machine in first public subnet
https://github.com/openshift/installer/pull/1121/commits/afa0b599b76163b08ea393bc36e94e6248042a8e had moved the bootstrap node to private
subnet based on https://github.com/openshift/installer/pull/1121#issuecomment-466563682, but we need the bootstrap node in public subnet to be able to ssh.

The bootstrap node is accesible on ssh again.
```console
$ ush core@18.215.154.240
Warning: Permanently added '18.215.154.240' (ECDSA) to the list of known hosts.
Red Hat CoreOS 4.0 Beta
WARNING: Direct SSH access to machines is not recommended.
This node has been annotated with machineconfiguration.openshift.io/ssh=accessed

---
This is the bootstrap node; it will be destroyed when the master is fully up.

The primary service is "bootkube.service". To watch its status, run e.g.

  journalctl -b -f -u bootkube.service
[core@ip-10-0-8-165 ~]$
```