data/aws: explicitly create the network interface for masters to speed up DNS
Today we create the network interface for master nodes implicitly as part
of the instance creation. We then use the IPs from those nodes to create
DNS records we use to set up etcd. This change creates the Network Interface
for masters explicitly instead of implicitly. Since this can be done before
the AMI copy is complete it means we can get DNS complete before we ever
launch a node. It looks like it saves at least 1m in the terraform time.

It might be more since DNS never will get an NXDOMAIN and cache that either.
I don't know about that yet