assets: no worker machinesets file for None
When the platform is none, there should be no worker machineset object created.
Currently when using the None platform, following file is created:

```console
$ ls -l dev-metal/openshift
total 20
-rw-r--r--. 1 adahiya adahiya  293 Mar  7 16:11 99_binding-discovery.yaml
-rw-r--r--. 1 adahiya adahiya  181 Mar  7 16:11 99_kubeadmin-password-secret.yaml
-rw-r--r--. 1 adahiya adahiya  330 Mar  7 16:11 99_openshift-cluster-api_cluster.yaml
-rw-r--r--. 1 adahiya adahiya   49 Mar  7 16:11 99_openshift-cluster-api_worker-machineset.yaml
-rw-r--r--. 1 adahiya adahiya 2689 Mar  7 16:11 99_openshift-cluster-api_worker-user-data-secret.yaml
```

The contents of the 99_openshift-cluster-api_worker-machineset.yaml are:
```console
$ cat dev-metal/openshift/99_openshift-cluster-api_worker-machineset.yaml
apiVersion: v1
items: []
kind: List
metadata: {}
```

`99_openshift-cluster-api_worker-machineset.yaml` should not be created when there are no machinesets to be created, either in 0 compute
required or None platform case.