libvirt CI Dockerfile fix:
Originally we installed nss_wrapper package from epel-testing,
I think because it wasn't available in epel repo (I'm not 100%sure)
We can now install from the stable epel repo, so no longer need the
epel-testing repo.  That's good, because epel-testing is no longer
configured in the base image (the build was failing until I removed
it, and I realized we no longer needed it).