cmd/openshift-install/create: Make waitForInstallComplete() more robust
If the cluster failed to completely initialize, we may still have a
router CA or console URL.  With this commit, errors in finalization
components are no longer immediately fatal; instead we hold any errors
but continue to try the later components.  We retain the first
component component error, and log any later errors.

To avoid waiting on a console URL when the cluster failed to
initialize, I've added a oneShot argument to waitForConsole.  When
waitForInitializedCluster failed, we set oneShot true and only make a
single attempt to resolve the console URL.  I've also shuffled around
the Until function to flatten some of its previous if/else blocks so I
only have to add single oneShot-cancel block.

Tangentially related, addRouterCAToClusterCA is idempotent, because
AppendCertsFromPEM wraps AddCert [1] and AddCert checks to avoid
duplicate certificates [2].

[1]: https://github.com/golang/go/blob/go1.12/src/crypto/x509/cert_pool.go#L144
[2]: https://github.com/golang/go/blob/go1.12/src/crypto/x509/cert_pool.go#L106-L109