pkg/destroy/aws: 2 minute lower bound for deletion period (was 10 seconds)
The 10-second bound is from 30504ed90 (pkg/destroy/aws: 10-second
lower bound for deletion period, 2019-03-02, #1352).  My two concerns,
discussed in that commit message, with raising this minimum were
slowing deletion time and the cluster self-healing.  But in our very
busy CI account (currently ~1200 clusters created, and destroyed again
within an hour or two, per day), we frequently hit AWS API rate
limits.  Looking at a recent run, it looks like my self-healing
concerns were unfounded:

  $ wget https://storage.googleapis.com/origin-ci-test/logs/release-openshift-origin-installer-e2e-aws-4.0/5703/artifacts/e2e-aws/installer/.openshift_install.log
  $ grep 'search for and delete matching resources by tag in us-east-1 matching.*kubernetes.io' .openshift_install.log | head -n2
  time="2019-03-15T04:12:46Z" level=debug msg="search for and delete matching resources by tag in us-east-1 matching aws.Filter{\"kubernetes.io/cluster/ci-op-8vcz3zix-5a633-xhhsv\":\"owned\"}"
  time="2019-03-15T04:13:38Z" level=debug msg="search for and delete matching resources by tag in us-east-1 matching aws.Filter{\"kubernetes.io/cluster/ci-op-8vcz3zix-5a633-xhhsv\":\"owned\"}"
  $ grep 'level=info .*arn:aws:iam' .openshift_install.log
  time="2019-03-15T04:12:47Z" level=info msg=Disassociated IAM instance profile="arn:aws:iam::460538899914:instance-profile/ci-op-8vcz3zix-5a633-xhhsv-master-profile" arn="arn:aws:ec2:us-east-1:460538899914:instance/i-07167482599a11b96" id=i-07167482599a11b96 name=ci-op-8vcz3zix-5a633-xhhsv-master-profile role=ci-op-8vcz3zix-5a633-xhhsv-master-role
  ...
  time="2019-03-15T04:13:28Z" level=info msg=Deleted arn="arn:aws:iam::460538899914:user/ci-op-8vcz3zix-5a633-openshift-machine-api-46hzq" id=ci-op-8vcz3zix-5a633-openshift-machine-api-46hzq

That means all of our IAM resources were removed during the first
deletion round (from 04:12:46Z to 04:13:38Z).  Without the IAM roles
and users, the cluster will be unable to create new resources needed
to self-heal, so we can make the limit as large as we like in that
respect.  At the moment, there's still a small hole in this because
the credentials operator doesn't create its own user/role for
credential creation, but we can close that hole if it turns out to be
a problem.

This two-minute limit is likely to increase teardown time, which
currently takes less than two minutes total during quiet periods:

  $ curl https://storage.googleapis.com/origin-ci-test/logs/release-openshift-origin-installer-e2e-aws-4.0/5660/artifacts/e2e-aws/installer/.openshift_install.log
  ...
  time="2019-03-14T08:39:48Z" level=debug msg="Built from commit 4b12088025778add89e95099eaeebb773c39436d"
  time="2019-03-14T08:39:48Z" level=debug msg="search for and delete matching resources by tag in us-east-1 matching aws.Filter{\"kubernetes.io/cluster/ci-op-s68jhhfx-5a633-45s9z\":\"owned\"}"
  ...
  time="2019-03-14T08:41:36Z" level=debug msg="search for untaggable resources"
  ...

but our dependency graph shouldn't be too deep, and waiting a few more
minutes for this unexciting task isn't the end of the world ;).