data/aws: create a public ssh security group
It is only used by the bootstrap node and becomes semi-vestigial after that
node is torn down. But a customer can use it to create nodes in the public
subnet which are accessable if they wish. They may also re-use this SG with
the existing master/workers if they want to enable ssh access to those nodes
after they do VPC peering/VPN setups