pkg/destroy/aws: Remove lastError value masking
In 200f0c907 (pkg/destroy/aws: Remove some lastError local-variable
masking, 2019-03-18, #1434), I removed some local lastError variables
that masked the function-level variable.  But Matthew points out that
we were still clobbering that function-level variable with the
loop-level value [1], so a successful loop iteration might silently
clear a previously-set lastError.  This commit goes through and uses
'err' for consistently for the loop-level variable.  When we have an
error, we log any previous lastError value before clobbering that
value with the new error.  It's up to the caller to decide how they
want to handle any final lastError; they can log it or not as they see
fit.

I've demoted the lastError logging from Info to Debug, because the
destroy logic usually uses debug for errors (e.g. DependencyViolation
errors), and I don't see a point to trying to classify errors as
expected or unexpected.

[1]: https://github.com/openshift/installer/pull/1434#discussion_r266954291