machines: add the authorized keys for a pool using a machine config
`cluster-config-v1` is being deprecated in favor of global configs [1] and Machine Config Operator needs to
drop using the `SSHKey` in install-config [2] to setup the `SSHAuthorizedKeys` for `core` user.

This pushes a machineconfig with the `SSHAuthorizedKeys` sourced from [2] for each machinepool, so that Machine Config Operator can drop
generating the machineconfig using the `cluster-config-v1` config map in the cluster.

[1]: https://github.com/openshift/installer/issues/680
[2]: https://godoc.org/github.com/openshift/installer/pkg/types#InstallConfig