pkg/asset/machines/worker: Default to MachineSets only where we need them
And not for additional zones.  This will keep us from consuming more
NAT gateways and associated EIPs than we need, once subsequent work
ties those resources to Machine(Set) consumption.  This means that
pools where replicas is unset or zero will receive no MachineSets
(unless the user explicitly configured zones for the pool), but
creating MachineSets later on is something we want to be easy anyway,
so I don't see a need to require the installer to inject a template
MachineSet into the cluster.

The "In each private subnet" -> "For each private subnet" change is
because the NAT gateways currently live *in* the public subnet, but
their purpose is to handle egress from machines in the private
subnets.