Set Kuryr as default NetworkType if platform supports it
This commit makes sure that the default NetworkType will be set to
"Kuryr" for OpenStack clouds that have Neuron trunk ports enabled and
are running Octavia loadbalancing service. For any other platform
OpenShiftSDN will be used. Please note that this does not override any
choice that might have been made by the deployer in the InstallConfig.

The detection and choice of the default value is done in
config.v1.Network's Generate() as at the point of
SetInstallConfigDefaults() we don't know anything about the target
platform capabilities.