image: Add a production "installer-artifacts" image for Mac binary
To ensure the payload is reproducible, we will include images for
both cli and installer commands for non linux platforms that allow
users to get the correct tools to install the release image. Since
the installer binary is not small and we don't want to take the cost
to pull it on each node when in use from hive, create a new image
installer-artifacts that layers on top of installer (inheriting the
default installer Linux binary) and places the darwin binary into
/usr/share/openshift/mac/openshift-install. This directory structure
is kept deliberately simple for end users - if we do in the future
need to deal with multi-arch concerns we'll do that at a higher level
and in practice neither 32bit nor arm will be "supported" as part
of the core distro yet.

The dockerfile matches the desired final form from the release team
(Dockerfile.rhel) and will be used in both CI and publishing.