upi/vsphere: add vsphere to cloudproviderconfig
Add the config for vsphere to cloudproviderconfig.

Add vsphere cloud creds to cloud-creds-secret.yaml.

Add fields to vsphere platform in installconfig to support cloudproviderconfig.

Set the enable_disk_uuid attribute to true for vsphere VMs as that
is required to ensure the VMDK properly mounts disks.

See https://vmware.github.io/vsphere-storage-for-kubernetes/documentation/existing.html
for details on the vsphere cloud config file.