docs/user/aws/images: Dia source for UPI arch diagram
Dia is from Stephen Cuppett, replacing the PNG he'd submitted via
39a926a918 (Adding initial user doc/guide & materials for UPI AWS
installation, 2019-03-12, #1408).

We aren't using the full file with all the layers yet, but I'm
building it anyway because folks without Dia may still want to look at
it ;).

SVGs generated with:

  $ dia --version
  Dia version 0.97.3, compiled 18:02:21 Feb 11 2017

relink-dia.py embeds the referenced icons in the SVG with def and use
[1,2] to avoid icon URIs like:

  file:///.../openshift/installer/docs/user/aws/images/./AWS-Architecture-Icons_SVG/Light-BG/_General%20AWS/AWS-General_AWS-Cloud_light-bg.svg

Ideally Dia would have a way to do this sort of thing automatically
with a command-line switch, but if it does, I can't find it.

[1]: https://developer.mozilla.org/en-US/docs/Web/SVG/Element/def
[2]: https://developer.mozilla.org/en-US/docs/Web/SVG/Element/use