pkg/asset/installconfig/aws: Extract regions from RHCOS metadata
Uncomment all of the regions in pkg/types/aws/validation (unwinding
that part of bd88157ab0, hack/build: Pin to RHCOS 400.7.20190306.0,
2019-03-12, #1407).  Instead, perform the "can we find an AMI for that
region?" check directly in pkg/asset/installconfig/aws when we're
building a list of regions for the wizard prompt.  With this change,
bumping the rhcos.json asset (via hack/update-rhcos-bootimage.py) will
automatically keep the wizard prompt's choices in sync with the
published AMIs.