FIXID BZ1686943: asset/store: prevent consuming targeted assets
When the user of the asset store fetches an asset, all of the assets
that the store thinks are on-disk are purged from the disk. When
the user is fetching multiple assets, the behavior should be that
all of the fetched assets are retained on disk. However, if one of
the fetched assets already existed on disk prior to the user
fetching the assets, then that asset will be purged from the disk
when the other assets are fetched.

For example, if the user runs the `create manifests` command,
then the Common Manifests and the OpenShift Manifests assets are
both fetched and written to disk. If the user then runs the
`create manifests` command again, then the fetch of the OpenShift
Manifests asset will cause the Common Manifests asset to be purged
from the disk.

The changes in this commit modify the Fetch function so that it
takes all of the assets being fetched at once instead of requiring
the user of the asset store to call Fetch for each asset. This allows
the asset store to make a more informed decision about which assets
should be retained on disk and which assets are eligible to be purged.

Additionally, the responsibility for persisting asset files to disk
has been given to the asset store. The asset store is responsible for
everything else related to the maintenance and lifetime of the asset
files. Why should it not be responsible for persisting the asset
files to disk as well?

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1686943