asset/metadata: save metadata file in asset state file
The Metadata asset uses a private field for its file. This
prevents the file from being saved when the asset state file
is generated. When `create ignition-configs` is called prior
to `create cluster`, then Metadata asset does not have a
file.

This does not present a problem to the end user yet,
because the Metadata asset, as a read-only asset, is not
consumed during the call to `create cluster`.