cmd/openshift-install/create: Drop addRouterCAToClusterCA
This was added in 40335776f6 (Append router CA to cluster CA in
kubeconfig, 2019-02-12, #1242), where the motivation was [1]:

  With users created with an identity-provider, OAuth flow goes
  through router and router-ca is not trusted. This prohibits the user
  from using oc login from command line, without manually appending
  the router ca to the cluster ca. The openshift-ingress-operator
  creates the router-CA and this is not available until the very end
  of an install.

  This PR adds the router CA from configmap router-ca -n
  openshift-config-managed to the kubeconfig
  certificate-authority-data. This allows an identity-provided user to
  oc login from the terminal.

But the admin kubeconfig is already authenticated, so folks shouldn't
be using 'oc login' with that kubeconfig.  For example, see dee6929c69
(Modify kubeadmin usage message, admins should not use kubeadmin via
CLI, 2019-04-01, #1513).  That leaves us with no use case for this
modification, and making the finish code watch-only sets us up to
rename away from the user-provided-infrastructure subcommand.

[1]: https://github.com/openshift/installer/pull/1242#issue-252678381