cmd/openshift-install/waitfor: Rename bootstrap-complete -> control-plane-ready
There is concern that "bootstrap" will be a less-familiar idea for
users, while "control-plane" will be more familiar.  This commit
pivots the subcommand name and its short description to use the latter
terminology.  We have decided to stick with bootstrap-complete
internally, and we consider the bootstrap-complete event to be part of
our internal API, so I've left all of the other code using
bootstrap-complete and derivatives (e.g. I'm still using
newWaitForBootstrapCompleteCmd).  This makes the rename as superficial
as possible.