upi/vsphere: remove extra users
The extra users were helpful when ramping up on booting rhcos
on vsphere. The boot process is more stable now so the extra
users are no longer needed. If problems do arise in the future
where having a user that can log in with a password is helpful,
that can easily be added manually to the ignition configs.