upi/vsphere: remove dependency on ipcalc
The cidr_to_ip.sh script used to reserve IP addresses uses
ipcalc to determine the subnet from the CIDR. The CIDR is
not used for anything else. These changes pass in the subnet
to the script rather than the CIDR so that ipcalc is not needed.
The motivation for this is that ipcalc is not available on
RHEL7, which is the base image for the image used in CI.