cmd/openshift-install/create: Log a more optimistic error for Progressing
And point at 'wait-for install-complete'.  This gives a better
experience for users that we time out because they're installing over
a slow network or something.

We might want to fail fast if we see a Failing, instead of waiting for
the full timeout.  But I'm leaving that alone for now; we can always
come back to it in follow-up work.