cmd: add gather bootstrap subcommand for gathering logs on bootstrap failure
This change adds a new subcommand to assists with debugging a failed
bootstrap by collecting logs from the cluster. Currently, it outputs the
shell commands to run in order to gather these logs. It can later be
extended to gather the logs itself. It is called automatically from the
create cluster subcommand on a failed bootstrap.