Drop AWS UPI control-plane Machines and compute MachineSets
Folks are free to opt-in to the machine API during a UPI flow, but
creating Machine(Set)s that match their host environment requires
matching a few properties (subnet, securityGroups, ...).  Our default
templates are unlikely to do that out of the box, so just remove them
with the standard flow.  Users who want to wade in can do so, and I've
adjusted our CloudFormation templates to set the same tags as our IPI
assets to make this easier.  But with the rm call, other folks don't
have to worry about broken Machine(Set)s in their cluster confusing
the machine API or other admins.

The awkward join syntax for subnet names is because YAML doesn't
support nesting !s [1]:

  You can't nest short form functions consecutively, so a pattern like
  !GetAZs !Ref is invalid.

Also fix a few unrelated nits, e.g. the unused VpcId property in
06_cluster_worker_node.yaml.

[1]: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-getavailabilityzones.html#w2ab1c21c24c36c17b8