docs/user: Drop compute access to etcd server from UPI docs
This was added out of an abundance of caution about how etcd metrics
were being scraped, because the Prometheus scrapers can run on compute
nodes.  We've now confirmed that etcd-metrics are being scraped from
port 9979, which is part of the existing "OpenShift reserved" and so
does not need a specific call-out.